users-service/requirements.txt
fastapi==0.100.0
pydantic==1.10.2
uvicorn==0.22.0
pymongo==4.7.0
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
pydantic[email]
python-jose==3.3.0